Fix RemoteCall & CustomForm

+++ ExtensionLibrary/Header/MC/Packet.hpp
@@ -15,8 +15,8 @@ namespace MC::Extensions
 
 		property unsigned Unk2 { unsigned get(); void set(unsigned value); };
 		property MC::PacketReliability ReliableOrdered {MC::PacketReliability get(); void set(MC::PacketReliability value); };
-		property uint64_t Unk24 { uint64_t get(); void set(uint64_t value); };
-		property uint64_t Unk40 { uint64_t get(); void set(uint64_t value); };
+		property uint32_t Unk24 { uint32_t get(); void set(uint32_t value); };
+		property uint32_t Unk40 { uint32_t get(); void set(uint32_t value); };
 		property uint32_t Incompressible { uint32_t get(); void set(uint32_t value); };
 	};
 }
